+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Version 2.x: in development
   [[ Bug report #75 ]](https://github.com/sybrenstuvel/flickrapi/pull/75)
   [[ Pull request #76 ]](https://github.com/sybrenstuvel/flickrapi/pull/76)
 - Configured Tox & Travis to test on Python 3.6 as well.
+- Timeout for API calls.
+  [[ Feature Request #27 ]](https://github.com/sybrenstuvel/flickrapi/issues/27)
 
 
 Version 2.2.1: released 2017-01-15